dataproject/data/requirements.txt
appdirs==1.4.3
asgiref==3.2.5
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
decorator==4.4.2
Django==3.0.4
dload==0.6
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.1
pandas==1.0.1
parso==0.6.2
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.4
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.6.1
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
regex==2020.2.20
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.11.2
